+++ finagle-integration/src/test/scala/com/twitter/finagle/integration/MetricsExpressionTest.scala
@@ -72,8 +72,9 @@ class MetricsExpressionTest extends AnyFunSuite with BeforeAndAfter {
         .build[Echo.MethodPerEndpoint](getAddress(thriftServer2), "client3")
         .echo("hi"))
 
-    // all expressions are created
-    assert(statsReceivers.forall { sr =>
+    // all expressions are created for thrift/thriftmux servers
+    // no deadline expressions for http2 server
+    assert(statsReceivers.dropRight(1).forall { sr =>
       sr.expressions.size == 5 &&
       sr.expressions.mapValues(_.name).values.toSet == keySet
     })