+++ CHANGELOG.rst
@@ -7,6 +7,11 @@ Note that ``PHAB_ID=#`` and ``RB_ID=#`` correspond to associated messages in com
 Unreleased
 ----------
 
+Runtime Behavior Changes
+~~~~~~~~~~~~~~~~~~~~~~~~
+
+* finagle-http: Remove DeadlineFilter from Server stack. ``PHAB_ID=D964479``
+
 Breaking API Changes
 ~~~~~~~~~~~~~~~~~~~~